+++ httperrors/httperrors.go
@@ -9,6 +9,19 @@ import (
 	"go.uber.org/zap"
 )
 
+var (
+	// AnError is an HTTPError instance useful for testing.  If the code does not care
+	// about HTTPError specifics, and only needs to return the HTTPError for example, this
+	// HTTPError should be used to make the test code more readable.
+	AnError HTTPError = &HTTPErrorImpl{
+		Status:      -1,
+		Err:         "TESTING ERROR",
+		Message:     "USE ONLY FOR TESTING",
+		GolangError: nil,
+		toLog:       true,
+	}
+)
+
 type HTTPError interface {
 	error