+++ persistence/models/dto/User.go
@@ -1,0 +1,9 @@
+package dto
+
+// Data Transfert Object Package
+
+// Login DTO
+type UserLoginDTO struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}