+++ router/router_test.go
@@ -13,8 +13,11 @@ import (
 func TestSetupRouter(t *testing.T) {
 	jsonController := middlewaresMocks.NewJSONController(t)
 	middlewareLogger := middlewaresMocks.NewMiddlewareLogger(t)
+	authenticationMiddleware := middlewaresMocks.NewAuthenticationMiddleware(t)
+
+	basicController := controllersMocks.NewBasicAuthentificationController(t)
 	informationController := controllersMocks.NewInformationController(t)
 	jsonController.On("Wrap", mock.Anything).Return(func(response http.ResponseWriter, request *http.Request) {})
-	router := SetupRouter(jsonController, middlewareLogger, informationController)
+	router := SetupRouter(jsonController, middlewareLogger, authenticationMiddleware, basicController, informationController)
 	assert.NotNil(t, router)
 }